Address thirty-third review pass: widen msg_id, make waiter installs session-atomic
Widen the wire `msg_id` from u16 to u32 (header 6 -> 8 bytes). My
content-based reuse detection from last round misses the likeliest case
outright: if the fragment the first message lost is exactly the one the
second supplies, the arriving index is not a duplicate, nothing looks wrong,
and the blend completes. Detection cannot close that — only making reuse
unreachable does. At 2^32 ids a sender cannot wrap back onto a live message
inside the reassembly timeout. The differing-payload check stays as a
backstop.

Cost: 2 bytes of payload per fragment, which at the 23-byte spec-minimum MTU
takes the budget from 14 to 12. That is a real throughput loss at worst-case
MTU and negligible at negotiated ones, and it buys correctness rather than
headroom.

Hold `advertise_lock` across the platform call, not just the generation
increment. Releasing early left a `stop_advertising` able to complete while
no attempt was installed — after which this future started a server the stop
had already returned from — and let two attempts claim generations in one
order and call Kotlin in the other.

Install Android operation waiters only while still owning the session.
Checking and then assigning were separate steps, so a stale handle could
overwrite the replacement session's `write_tx`; Kotlin then rejected the
stale operation, that rejection removed the sender, and the replacement's
own callback found no waiter and left its future pending forever. Ownership
check and waiter install now share one critical section.

Remove subscription state on every unsuccessful exit. `onSubscribed` only
removes `subscribe_tx`, so a rejected, failed or cancelled subscribe left
`notify_tx` holding a sender whose receiver was gone — growing per distinct
UUID until disconnect, and clobbering a live route when a retry for an
existing UUID failed. A guard now removes both, conditional on request id so
it cannot strand a retry.

Reject mock writes to unknown or non-writable characteristics. The mock
accepted any UUID and ignored the advertised permissions, so a mistyped
characteristic or a permissions mistake passed here and would fail only on a
device.

Verified: cargo test --workspace (62 passed, 8 repeat full-suite runs, no
flakes); clippy -D warnings clean on desktop, x86_64-linux-android, and the
plugin crate on aarch64-linux-android; npx tauri android build (APK + AAB).

+++ ble-gatt/src/datagram/mod.rs
@@ -293,7 +293,10 @@ pub struct DatagramChannel {
     /// would be the first thing discarded — the failure would silence its
     /// own alarm.
     overflow: Arc<OverflowSignal>,
-    next_msg_id: u16,
+    /// Rolling message id. 32-bit so it cannot wrap back onto an id whose
+    /// message is still being reassembled — see `fragment`'s header docs for
+    /// why 16 bits was not merely tight but unsound.
+    next_msg_id: u32,
     max_message_len: usize,
     /// Fixed at construction from the negotiated MTU. BLE does not
     /// renegotiate mid-session in practice, so this is not refreshed.
